Delete AGENTS.md parsers, keep the migrate command's copy
parse_agents_config is gone from arkavo-protocol (a breaking API change
for external consumers of the crate, and of arkavo-agent's re-export —
the point of S6) along with the dead parse_workspace_paths/WorkspacePaths.
parse_runtime_config stays: sequence_integrity_test.rs still calls it.

The CLI's copy moves out of agent.rs into commands/kit/legacy_agents_md.rs
(split from its YAML-property helpers to stay under the file-size cap),
renamed to parse_legacy_agents_md so the grep gate on the old name is
clean, and kept alive only for `arkavo kit migrate-from-agents-md`.

Also fixes two doc comments (types.rs's AgentConfig{Get,Update} content
fields, chat_session.rs's system_prompt) that had gone stale describing
"AGENTS.md content" when the RPC payload and agent purpose have both
been kit-sourced since the S6 cutover.
